fix the BO benchmarks + move functions to waf_tools/benchmarks.py
